Initial draft to port DLUX to modularity. All modules converted, and also add a css injector library. We can load Css dynamically HOW IT WORK
Angular modules are cut into small requirejs module. The module definition and his configuration will be in a file named <module_name>.module.js . The controllers will be in a <module_name>.controller.js.  It's the same thing for the directive, services (factories), providers and filters.

The web page has one physical dependency; requirejs. RequireJS will look for the file main.js. This file is use so that it can configure himself and to load all vendor dependencies. Next, a call is send to the App.js. At this state, all dependencies for the app.js are hardcoded but they will be provided by a web application server. The web server will look for all file who finish by *.module.js and send it to the application.

When a requirejs module is loaded, it will look for his dependencies and load them if they aren't already loaded.  When everything is loaded correctly, the angular bootstrap is done.  After most of the job is done in the configuration method of each angular module. First we create a register for our services, controllers, filter and directive. Because at this state the app is already bootstrapped so to be able to call them we need to register them. Next, we set the url route and use the core helper if needed. At the end, we need to load the controller otherwise  it won't be loaded. So we use ocLazyLoader to load them when we need it.

Base App modules
-- Core => Contains all helpers to populate the TopBar and the NavBar.
-- Layout => Takes everything in the helpers and adds it to the html page
-- Navigation => Creates the menu.

Change-Id: I67129bdf3f1f61e2de74fa97f2617918a9405bc0
Signed-off-by: Maxime Millette-Coulombe <mmcoulombe@inocybe.ca>