Fixing the basic login issue
Dlux-ui fails to login using admin/admin due to the incorrect base-url configuration in development.json.
It is important to change the defaults after the authentication works end-to-end and let the application
decide on which Base-url to use.

Changing it to localhost:8181 and trying to access FlowProgrammer nb/v2 is not going to succeed.

Change-Id: I672ad4d201e191360c935eac6cc39a65228bfe95
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>