Repair broken tests -DO NOT MERGE-
All tests look running fine except those with state transition in the node module.
The error is when the state is loading, he also load his module. Because the state
is a promise, we don't see it fail. I try to solve http request error caused by
the unit test env. (The module try to get app/module but its base/app/module)

Change-Id: I7f8340821ec0993d0141f1cc3878ffb4d58d58c4
Signed-off-by: Maxime Millette-Coulombe <mmcoulombe@inocybe.com>