Use grunt-maven-plugin to build dlux-web
With this plugin the user now only has a single requirement on npm being
installed prior to building. This plugin handles getting "bower" and
"grunt" as part of the Maven build. This makes building dlux-web a
little more portable.

It uses maven-resources-plugin to copy the generated-resources in
target-grunt over to target/ and maven-antrun-plugin to delete
extronious directories created by the grunt-maven-plugin before finally
creating the jar.

Change-Id: I8f232301914fdbdd503363aafcab862fc982ee68
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>