Update patch set 1
Patch Set 1:

As far as I can tell this patch creates a final jar file that is equivalent to the jar file created before this patch is applied. I verified this by comparing the old jar file and the new jar file created by this patch.

I did not however verify that the jar produced the correct final code as I'm not too familiar with npm/bower/grunt.

Patch-set: 1
Label: Verified=0