Use frontend-maven-plugin to build dlux-web
With the frontend-maven-plugin we can eliminate the need for the user to
have npm / bower / grunt installed. This plugin bootstraps itself by
downloading node / npm then we use the exec-maven-plugin to run bower
to install dependencies. Finally we run another phase with the
frontend-maven-plugin to run Grunt to complete the build.

Change-Id: I8f232301914fdbdd503363aafcab862fc982ee68
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>