Update patch set 9
Patch Set 9:

> > Do we need it in master and stable/lithium?
 > 
 > This is very likely.  The code in stable/lithium has changed
 > considerably since stable/helium, so its not a straight cherry
 > pick.

It looks like a bunch of code was ripped and replaced with Restangular.  I am not an angular expert;  do you know someone who could make this change in Li?  If not, let me know and I will figure it out, albeit slower than the experts can.

Patch-set: 9