Update patch set 1
Patch Set 1:

The point of this first patch is not the test functionalities of the authentication module. It make it pass the loading process of the authentication module. I divide  it from usual test case cause i thought it was a distinct issue.

Patch-set: 1
Reviewer: Gerrit User 1951 <1951@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0