Remove DLUX distribution because it causes cyclic build dependency
The DLUX distribution pulls in l2switch which depends on openflowplugin
however openflowplugin depends on dlux. This causes an unresolvable
build cycle every time we version bump post release and is incredibly
painful to work around. This patch removes the DLUX Distribution to
resolve this cycle.

There's been a suggestion on the mailing list for DLUX to use the
integration-distribution for their testing rather than building their
own custom distro.

See:
https://lists.opendaylight.org/pipermail/release/2015-December/004743.html

Change-Id: I1b61406a4f2c3e24d42a78c69f2943d8f775c16d
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>