Freeze version of external dependencies
Recently, an error during build time was introduced because one external
dlux dependencies has been updated. In the version resolving tool, the
extracted library was always the latest. To avoid other similar issues,
all libraries must need to be resolved according to the public API used
and not using "latest" or "master". According to semantic versioning,
those values allow the versioning tool to fetch library version with a
different API.

Change-Id: Iffb7dab2ab8a7bf196bd7a0252d69ba7e5576cc0
Signed-off-by: Maxime Millette-Coulombe <mmcoulombe@inocybe.com>