Update patch set 2
Patch Set 2:

I'm concerned here. Will this reintroduce a cyclic dependency between repos?

I recall we purposely removed distribution-karaf from dlux in the past because of cyclic dependencies causing it to be impossible to build dlux and another project. Also note that just because it passes a SNAPSHOT build does not mean it does not have this issue because Maven falls back to Nexus and can successfully download the missing dependency that it shouldn't be able to resolve.

Patch-set: 2
Label: Verified=0