Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > there's something in npm-land :(
> > 
> > the problem is that it is working on my computer both Windows and Ubuntu.
> 
> It fails on mine, too. I tried upgrading frontend-plugin, but that does not help :(
> 
> @zxiiro I seem to remember you casting some magic here ... any ideas?

I proposed this patch (https://git.opendaylight.org/gerrit/c/dlux/+/76402) a year ago that at the time fixed the issue in the stable/oxygen branch. It was abandoned on the master branch because we decided to archive DLUX. Let me do a local build with the new version and see if it works.

Patch-set: 3